Split projects page tests based on visibility
